requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
python-docx>=1.1.0
openai>=1.12.0
pydantic>=2.9.0
python-dotenv>=1.0.0